flamenco, fuzz: rename vm test API file
